% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_drift.R
\name{DriftToInds}
\alias{DriftToInds}
\title{Sample simulated individuals from stream sites.}
\usage{
DriftToInds(x, xs, sites, n)
}
\arguments{
\item{x}{Drift output. The output of a stream.drift function call.}

\item{xs}{data.frame. xs data, as produced by \code{\link{Edge.to.Matrix}}.}

\item{sites}{data.frame. A list of the sampling sites, where the first column is the branch ID and the second is the sampling position in miles downstream from branch start. Allele frequencies will be interpolated for sites between simulated points.}

\item{n}{numeric vector. The number of individuals to sample per site, in the same order as the sites data.frame.}
}
\description{
Samples individuals from any number of sites based on the simulated minor allele frequencies at those sites.
Genotypes are in the format 0, 1, or 2, signifying the counts of the frequency of the allele initially given to
\code{\link{stream.drift}}.
}
