% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_drift.R
\name{IndsToSFS}
\alias{IndsToSFS}
\title{Generate a 1 or 2D site frequency spectrum from simulated individuals.}
\usage{
IndsToSFS(genotypes, sites, fold = F)
}
\description{
Generates a 1 or 2D site frequency spectrum from individuals simulated via \code{\link{DriftToInds}}.
If not folded, assumes that the tracked allele (that for which the frequency was initially provided to \code{\link{stream.drift}})
is the derived allele.
}
