% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_drift.R
\name{plot.stream.drift}
\alias{plot.stream.drift}
\title{Plot average allele frequency changes across a river map}
\usage{
\method{plot}{stream.drift}(drift.dat, edge.dat, xs, viridis.option = "viridis")
}
\arguments{
\item{drift.dat}{list. List containing genetic drift data produced by \code{\link{stream.drift}}.}

\item{edge.dat}{list. Edge and map data produced by \code{\link{GIS.to.Edge}}.}

\item{xs}{data.frame. xs data, as produced by \code{\link{Edge.to.Matrix}}.}

\item{viridis.option}{character, default "viridis". The name of the viridis option to use for plotting, for details
see documentation for \code{\link[ggplot2]{scale_color_viridis_c}}.}
}
\value{
A list containing \itemize{\item{delta_plot: } A ggplot map of average allele frequency change. \item{fixed_plot: } A ggplot map of the percentage of fixed alleles. \item{data: } The raw data used to produce the plot.}.
}
\description{
Create plots of allele frequency change across a river map using drift data from \code{\link{stream.drift}},
and edge data.frames and map data from \code{\link{GIS.to.edge}}. Maps are produced using \code{\link[ggplot2]{ggplot}},
and are modifiable as typical for those objects.
}
\author{
William Hemstrom
}
